goodbye codify model saving in safe_tensors format
