--- refact_models/__init__.py
@@ -1,1 +1,0 @@
-from refact_models.codify_model import CodifyModel
