+++ self_hosting_machinery/inference/__init__.py
@@ -1,4 +1,3 @@
 from self_hosting_machinery.inference.inference_base import modload
 from self_hosting_machinery.inference.inference_base import InferenceBase
-from self_hosting_machinery.inference.inference_legacy import InferenceLegacy
 from self_hosting_machinery.inference.inference_hf import InferenceHF
